miner: Support remote sealing service
